fix search issue
